fix(git): make "j" behave properly upon showing pager
The character you might not see here at the start of the pattern is
CTRL-K, which behaves as described in the `less` man page:

    Certain characters are special if entered at the beginning of the
    pattern; they modify the type of search rather than become part of
    the pattern:

    ^K     Highlight any text which matches the pattern  on the current
           screen,  but don't move to the first match (KEEP current
           position).

Without this, you land in the pager looking at the output of a command
like `git log -p`, the first "commit" is not highlighted, presumably
because you were already "there", which means that the search pattern
actually matches the first "diff", and that's where you "are". When you
hit "j" you wind up scrolling down from there, which isn't at all what
you expect.

CTRL-K prevents this unwanted behavior.

Note that because it "doesn't move to the first match", hitting "n" also
works as expected (ie. you jump down to the first "diff" when you hit
"n").