feat(git): add "menos" tool to improve paging
This fixes the edge case where you run `git diff` in a repo with no
changes. Before all our recent tinkering with paging, Git would emit
nothing here, and it would go through `less -iFMRSX` and `less` would
immediately exit because the output doesn't fill the entire screen (ie.
`-F`).

But now that we are using `--pattern` and `+r`, this case blanks the
screen.

The workaround applied here is to make a little Ruby script that stands
in for `less`. It tries to read a little bit of stdin into a buffer; if
there is nothing to read, it calls `less` but strips out the problematic
`--pattern` and `+r` arguments. Otherwise, it forks `less` normally and
let's it get to work.

It is called `menos`, which is Spanish for "less".