refactor(menos): go back to buffering
Because I have found that any use of `--pattern` at all causes `less`
(not just `menos`, but `less` itself when used directly too) to clear
the entire screen even when `-F` is in effect; so you might wind up with
a command like `git log -2` or `git diff` producing only a few lines of
output, but it clears the entire screen which kind of ruins the desired
effect of `-F`.

So we check to see if we can fit all output on one screen, and if we
can, we skip the whole `--pattern` thing.