chore(fig): port automator "role"
Doesn't actually do anything, but maybe I will be able to, er, automate
this in the future.

Note that because I don't have this aspect listed in the project.json,
even if you run `./install automator`, it won't even print the "skip"
message because it does literally nothing. Consider this aspect
aspirational and a placeholder.

Also note the bad filename of the old file; I don't think I had a good
reason for this -- it was almost certainly a mistake.