fix(karabiner): avoid chicken-and-egg dependency on node
Don't want to have to rely on "node" having been previously installed
and available in the path. Use the `bin/node` wrapper script instead,
which we know must be present in order for us to be running at all.

This means that even on the first run (before `$PATH` includes an active
installed entry under `~/n`) that we can use Node here.