feat(vim): add Flow support
Seeing as I am still installing the flow-bin package.

Note that I am putting it at the end because I want the TS configuration
to take precedence, although I think it will anyway because of the
defaults which let tsserver handle projects with "package.json" (as well
as "tsconfig.json") whereas they only activate Flow in the presence of
".flowconfig").