refactor(zsh): drop "S" from LESS environment variable
From "man less":

    -S or --chop-long-lines
          Causes lines longer than the screen width to be chopped (truncated) rather than
          wrapped.  That is, the portion of a long line that does not fit in  the  screen
          width  is  not  shown.  The default is to wrap long lines; that is, display the
          remainder on the next line.

I want this in Git commands, but generally not for general `less` usage.
For example, if I do `less README.md`, I want the lines to wrap so that
I can read it without side-scrolling (Markdown files are often not
hard-wrapped). But if I do `git log -p` I don't want wrapping.

So, move the "S" out of `LESS` and into Git's `core.pager` config. Net
effect is that Git behaves the same as before because it is doing the
equivalent of:

    echo git-command-output | LESS=-iFMRX less -S

instead of:

    echo git-command-output | LESS=iFMRSX less

but the end result is the same.