fix(git): avoid annoying breakage of starting "/" in Git pager output
Originally this commit was going to be a different hack to remedy an
annoying problem with `less --pattern=anything +r`, which breaks an
initial "/" on entering the pager (causes it to print "Pattern not found
(press RETURN)").

So, I had code in here to look at our buffer and:

- If there is a match, we know we don't need the "+r", so strip it:
  problem solved!
- If there is no match, we just delete both "--pattern" and "+r".

But I realized that if I just remove all of that (no "--pattern", no
"+r" in our Git "pager.<cmd>" config, and no special handling inside
"menos"), we can rely on our other hack, the one that manipulates the
LESSHISTFILE, to do the trick for us.

Bonus: this stops `less` from highlighting the search pattern as soon as
you hit "j", "k".

The ugly bit: "menos" looks like a generic wrapper, but it isn't really.
It (still) assumes a particular "pattern" is being passed (hence our check for
"commit|diff") and needs to be kept in sync with whatever is in
".gitconfig". A "proper" fix here would make a regexp from the pattern,
but that is made trickier due to the possibility of special characters
like C-k... I still might do that as a follow-up.