fix(vim): make sure LSP client can find root
Opening a file in "fig/" was causing the root to be misidentified as
"fig/" because of "fig/package.json" being a root marker. This would
lead to spurious errors (eg. inability to see root tsconfig.json, which
would cause things like "target" to be misdetected).

I think we don't need the package.json at edit time (under "src/"), but
we do need it at runtime (under "lib/") so let's just move it.