chore(ranger): swap "zf" and "f" ("filter" and "find")
"find" is way too potent to be the shorter option. It will run the
selection as soon as it finds an unambiguous match. ie. if you type "z"
and there is a file in the current directly that uniquely has it, it
will open it. I'd rather use "f" for "filter", and I don't think I'll
ever use "zf" to be honest.