chore(vim): document some bogus attempts at improving popup windows
Too tired to sort this out now, so leaving some evidence for later
follow-up.

When we show the popup, it is #303030 (basically, same as NormalNC,
ColorColumn etc).  When we focus into it, it gets brighter (#505050,
basically, same as Visual, NormalFloat, Pmenu etc) and the window you
came from is now non-current so goes to NormalNC.

This is all a bit distracting. I can't for the life of me make up my
mind how I'd like this to work. This effectively inverts the colors when
you switch focus, which seems a bit too extreme:

    highlight clear NormalFloat
    highlight link NormalFloat Normal

The commented out code here represents a couple of failed experiments at
setting 'winhighlight' on window creation. I think these are getting
clobbers by my other autocmd fanciness... Will have to look into this
again with fresh eyes tomorrow or later.