refactor(menos): dramatically simplify, removing buffering
If you look at the history of this script you'll see that it was built
up incrementally to address a bunch of specific issues: intercepting
`--pattern`/`+r` to catch redraw issues, handling interrupts specially,
responding specially if output fits on one screen etc.

Today, I noticed one remaining issue: when running a command like
`git log -Gblah` on a huge repo that might take a while to produce a
screenful of results, no visible progress is apparent until the whole
screen is buffered. And then the subsequent buffering after the first
screenful probably doesn't help much either.

Looking at all this, I realized I no longer need most of the code in the
script because it was the product of iterative mutations that don't make
sense any more. We can just fork and run `less` after manipulating the
history file.

With a reduced test case I was able to determine that there is still
some unwanted buffering going on even in the case when Git can start
producing output immediately (about four seconds), because I still see a
difference between:

    # Uses `less`: prints first lines immediately, 4 seconds overall.
    git --paginate log -SFieldBase.scss --since=2.weeks.ago

    # Uses nothing: prints first lines immediately, 4 seconds overall.
    git --no-pager log -SFieldBase.scss --since=2.weeks.ago

    # Uses `menos`: prints nothing for 4 seconds, then prints all at once.
    git log -SFieldBase.scss --since=2.weeks.ago

    # No difference.
    env GIT_FLUSH=1 !!

But even if `menos` is reduced to:

    #!/usr/bin/env ruby

    exec 'less', *ARGV

you still see the lag. I believe it's because it is going through the
`diff-highlight` Perl script, which is doing its own buffering. The
real fix, then, is to fix that script to flush immediately. But in the
meantime, simplifying the script here is a win, so going to do that.