feat(vim): highlight nonbreaking space in Goyo buffers
Trailing whitespace still gets highlighted (once you leave the line), so
no worries about that, but nonbreaking spaces are notoriously frequent
in the kinds of buffers that I use Goyo in (ie. blog posts etc, where it
is super easy to type "option-hyphen" to get an emdash, and then
accidentally follow it with "option-space", which produces a nonbreaking
space), and worst of all, the error isn't visible in the Git diff
output.