feat(vim): set up TextYankPost
This is going to be a great little speed-up. Instead of visually
selecting things and then yanking them (to be sure you got he right
thing), you can just blindly yank with the same level of confidence
because if you get it wrong, you'll see it in the highlighting.

Thanks to: https://github.com/neovim/neovim/pull/12279

Screenshot: https://twitter.com/wincent/status/1264340870322675713