feat(karabiner): add "Launcher" mode
Hold "o" and a shortcut key to open or switch to an app; examples: "o" +
"t" (terminal), "o" + "b" (browser etc).

Inspired by the "Hyper Mode" section in this blog post:

https://medium.com/@caulfieldOwen/turn-your-keyboard-into-a-text-editing-rocket-1514d8474d2d

I realized that I rely pretty heavily on Command-Tab to switch
applications and this could be a nice way to eliminate some of the
inefficiency (sometimes you have to cycle through multiple steps to get
to a target, and this can serve as a shortcut).

Right-option seems to be a popular choice for the "hyper" key but I find
that I use it often enough as a real modifier that I'd probably get
confused if I tried to repurpose it.