chore(karabiner): disable launcher layer for now
Because it is driving me crazy with false positives, despite the
tweaks in a9fb5e7738f1a3751a25. eg. very often when I type a word like
"function", it detects "nc" as a simultaneous keypress and sends me to
Chrome.

I don't know why this is happening so often with this layer, but I never
get bothered with unwanted triggers in SpaceFN. No time to delve deeper
into this now, so just deactivating the whole thing for now.