feat(zsh): run `portool setup` as part of daily build
I get the impression that the "osgi/configs" directory gets reset on
every `ant all` (or perhaps at some other time), because looking there
today, I can't see the config files I added yesterday (and the only
thing I've done so far today is a `portool daily`).

So, let's just do this every time. I think it should be safe. Note that
I am running it for both the "master" and "portal" targets, which means
it will actually run twice when doing the standard "daily" (which does
both of those); it's idempotent, so harmless enough to do it twice even
though it is "global" in some aspects (ie. it sets up properties files
in both "master" and "portal"... although I should probably change
that).