perf(vim): avoid repeated focus and blur applications
We need to subscribe to a bunch of autocmds in order to cover all
scenarios, but this has the unfortunate side effect that we may repeat
our blur and focus side-effects. Some examples:

-   Running `vim` or `vim something.txt`:

    BufEnter
    VimEnter
    FocusGained

-   Command-Tab and back:

    FocusLost
    FocusGained

-   Splitting with `:sp`:

    WinLeave
    WinEnter

-   Splitting with `:sp something.txt`:

    WinLeave
    WinEnter
    BufEnter

-   Opening a new file with `:e something.txt`:

    BufEnter

-   Opening with Command-T:

-   Moving between splits:

    WinLeave
    WinEnter

-   Moving between tabs:

    WinLeave
    WinEnter
    BufEnter

So, not an exhaustive survey, but you can see that we have a problem: up
to three "focus"-triggering events for one actual action, but not always
the same ones:

    Open          Tabbing       `:sp`     `:sp x`   `:e x`

    BufEnter      FocusGained   WinEnter  WinEnter  BufEnter
    VimEnter                              BufEnter
    FocusGained

The situation is a bit better for "blur"-triggering events, which only
seem to happen in isolation:

    FocusLost     WinLeave

Whatever, the solution is to maintain an explicit "focused" flag in a
per-window variable.