feat(zsh): make `portool run bundle` work on older branches
Just noticed this testing a bundle on an older DXP branch (7.1.x), that
it is using an older version of tomcat (9.0.33 vs 9.0.37), so the
hard-coded `$TOMCAT` won't work. Let's autodetect it instead.

The other thing it's going to choke on is the MySQL driver name in the
portal-ext.properties; that one is going to be harder to autodetect,
so I'm just leaving this here as a reminder for posterity: on old
branches you are going to want `com.mysql.cj.jdbc.Driver` instead of
`com.mysql.jdbc.Driver`,and will have to edit it by hand.