fix(vim): handle edge case in pinnacle.decorate()
We can't assume that the term we're looking for will be proceeded by
whitespace because it could conceivably be the first item in the
string. If it is, the frontier check (`%f[%S]`) will fail to match,
because we are not transitioning from non-whitespace to whitespace). On
the other hand, the `%f[%a]` (word-boundary) check will work.

Sample REPL session to "prove" this:

  > =("thing=true gui=crap other=not"):match('^(.*)' .. '%f[%S](gui=%S+)' .. '(.*)$')
  thing=true      gui=crap         other=not

  > =("gui=crap other=not"):match('^(.*)' .. '%f[%S](gui=%S+)' .. '(.*)$')
  nil

  > =("gui=crap other=not"):match('^(.*)' .. '%f[%a](gui=%S+)' .. '(.*)$')
          gui=crap         other=not

  > =("thing=true gui=crap other=not"):match('^(.*)' .. '%f[%a](gui=%S+)' .. '(.*)$')
  thing=true      gui=crap         other=not

ie.

- with `%S`, not at beginning of string: works
- with `%S`, at beginning of string: returns nil
- with `%a`, at beginning of string: works
- with `%a`, not at beginning of string: works