fix(vim): get rid of unwanted ".." in statusline filename
Given a file like:

    src/scripts/webpack.js

and a string in the file like:

    '../utils/mergeBabelLoaderOptions'

if I hit "gf" on that string, Vim opens this file:

    src/utils/mergeBabelLoaderOptions.js

but shows it in the status line as:

    src/scripts/../utils/mergeBabelLoaderOptions.js

Then the call to `expand('%:h')` turns it into an absolute path
something like:

    $HOME/$BLAH/src/scripts/../utils

and `fnamemodify(file, ':~:.')` preserves the '..'. You can observe the
same behavior if you pass just ':~' or ':.' in isolation. The solution
is to add another ':p' to the start of the list of modifiers.

In short, the behavior of `expand()` as we were using it is undesirable
for our use case. We could also have passed '%:p:h` to `expand()` to
counteract it, but that would have made our check against '.' cease to
work, so changing it like we have here seems preferable.