fix(vim): handle another statusline "edge" case
Looks like my comment in the parent commit wasn't right:

> In short, the behavior of `expand()` as we were using it is undesirable
> for our use case. We could also have passed '%:p:h` to `expand()` to
> counteract it, but that would have made our check against '.' cease to
> work, so changing it like we have here seems preferable.

After that change, opening a file in the current directory with, say,
`vim filename.ext` our `expand()` call would convert it to:

    .

and we'd immediately return and all would be rosy. BUT if we got to that
file after passing through a vim-dirvish buffer, the `expand()` call
would yield:

    /absolute/path/to/the/files/directory

And then `fnamemodify()` would turn it into '', onto which we would
append '/', producing this:

    /

which is not what we want.