fix(vim): fix blur edge case
So I'm effectively reverting 4d22b87fd00e34d3bf2d3ee72b0d26d88db03773
because I realized it is a bit more complicated than all that.

If you run `vim some-file.js` everything is great, but if you run `vim`
and thus enter an empty buffer, and then open a file with `:e`, say,
then blurring doesn't work properly on your next `:vs` operation or
similar.

Why is this? Because when you start from an empty buffer and then open
a file, our `ownsyntax()` function gets called and ends up setting
`wincent_ownsyntax` to `false` on the window before actually entering
the buffer. Then when you spilt, `wincent_ownsyntax` is already `false`
and so we think we don't have to do anything. We can avoid this by only
setting `wincent_ownsyntax` to `false` when there is no `filetype`; that
is, we skip doing it on the empty buffer, and everything is rosy.