feat(vim): set up clangd LSP
Note that on macOS this is not added to the PATH by default (because it
would clash with Apple's own LLVM install), so I am specifying the full
path here.

Needed to update Neovim in order to get this to work without errors (eg.
`brew unlink neovim` and `brew install --force --HEAD neovim`). Followed
by:

  pip3 install --upgrade pynvim
  pip3 install redis commandt.score

and `:UpdateRemotePlugins` (not strictly sure if I need that or not, but
just in case... `:checkhealth` is clean after this).