feat(zsh): expand aliases after "sudo"
Every now and again I'll try to "ll" a folder which I don't have
permission to see as a normal user, so I have to do "sudo ls -laF" when
I'd really rather just do "sudo ll" and preserve muscle memory.

I learned today this trick for Bash:

https://twitter.com/wincent/status/1333036294440620036

where you terminate an alias with a space in order to continue alias
expansion for the following word.

It happens to work on Zsh too.