fix(vim): correct another problem with statuslines in quickfix windows
Previous attempted fix (in fa604748) was incomplete because this
variable is used in two places. Last commit fixed it when used in the
context of Lua, but broke it in the context of a Vimscript `:execute
setlocal` command. I believe now though I've got it working properly in
both places.