feat(aur): add role for installing from Arch User Repository
Convenient way to get neovim-nightly, for example.

The one gotcha, neovim-nightly doesn't have Ruby baked in, so will need
to figure out something for Command-T (am experimenting with rewriting
it in Lua, but that is still a way off).

Note this required the addition of the `raw` option to the `command`
operation, because otherwise its helpful attempts to normalize path
arguments corrupt URLs like "https://aur.archlinux.org/yay.git" to
"http:/aur.archlinux.org/yay.git" (note the single slash), causing the
clone to fail. Probably reason enough to add a `git` (or `get`)
operation in the future, if I ever need to clone more repos.