fix(vim): always call clipper#set_invocation() no matter what
Forgetting my own advice here. In the vim-clipper docs it says:

    Note: This is a function and not a setting for security reasons; only the
    first call to |clipper#set_invocation()| will have any effect. In this way,
    you can place the call near the start of your |vimrc| without having to
    worry that a malicious plug-in will later overwrite it with something
    nefarious (important because whatever is passed to this function will be
    blindly executed with the |system()| function).

    As such, even if you don't want to use this feature, it is a good
    idea to lock it down by calling `clipper#set_invocation('')` in your
    |vimrc|.

So I used to have it (more or less) like that, in the sense that I was
doing it in an "after" plugin, but I removed that in 5ad58f66. I'm
not moving it to my vimrc because it's a bit painful; you'd have to load
vim-clipper before any other plugins, then call `set_invocation()`, then
load the rest, and I am not sure it's worth it.