fix(dotfiles): partially fix root status not showing up in prompt
Problem - on macOS /etc/sudoers has:

    Defaults env_keep += "HOME"

but Arch Linux does not. The Arch default is more secure, so I am loathe
to change it, but it means that we get different behavior:

- on macOS, `sudo -s` dumps us in a zsh shell with `$HOME` set to the
  originating user's HOME. As such, `~/.zshrc` gets sourced and the
  prompt updates (shows "root" and correct SHLVL-based indicator).
- on Arch, `sudo -s` sets `$HOME` to `/root` and user `~/.zshrc` is not
  sourced; so no "root" in prompt and no SHLVL updates. `PS1` gets
  inherited, but it is static and reflects the immediate prior value.
  `sudo -Es` doesn't suffer this problem, but it's not great either (ie.
  it basically switches the behavior to nearly match macOS, perhaps a
  bit looser even, and so it laxens security with the added drawback
  that you have to remember to do it).

So, the partial fix implemented here is to avoid the dynamic `EUID`
check which happens only when the `.zshrc` is sourced, in favor of a
static/declarative approach that is inherited and works every time a new
shell is spawned, even when `.zshrc` is not re-sourced.

It's partial because the SHLVL logic is still stale in these cases. I
believe I can fix it somewhat but it will be fiddly: `man zshmisc` notes
how you can nest expressions, and there are markers for checking the
shell level. So I could "unroll" a loop and have the markers update at
least up to a certain level; don't think I can do any of the conditional
logic though for detecting when I'm in tmux or not... not sure if it is
worth it, so going to think about it.