feat: make Return serve as Control when chorded, return when tapped
Geez, this took altogether too long to figure out. Tried for about a day
with xmodmap + xcape, then realized that xkb is the new hotness and
xmodmap is basically legacy/deprecated; couldn't get it to work though
in several hours of trying. Finally settled on Interception Tools, which
seem to work fine.

- https://gitlab.com/interception/linux/tools
- https://gitlab.com/interception/linux/plugins/dual-function-keys