feat(dotfiles): provide sample xkeysnail config
As noted in the comment though, I don't think I can use this with the
udevmon service as they both try to grab the device.

I might be able to implement something similar-ish using Interception
tools though (basically add something that checks X to see the frontmost
application; not sure whether that will be prohibitively slow or not
though...).