feat(interception): add seed version of mac2linux
Just a POC to show that it's possible to set up
non-application-specific overrides at the udevmon level. Due to
shortcomings of xkeysnail, I'll milk this particular approach for all
it's worth, but I will almost certainly need to keep xkeysnail around
for some application-specific things.

This really is just a POC: it remaps a single key combination, just to
show that the mechanism works -- Alt-L is remapped to Ctrl-L (can't do
this in xkeysnail because Chrome somehow intercepts the key events
before xkeysnail gets them).

Next steps will be to generalize this to more overrides. I imagine I can
use this for all manner of things, including faking the macOS-style dead
keys. If the list grows too long (and it will), I will probably make it
possible to configure this via a YAML config file, like some of the
other Interception Tools stuff does.