feat(interception): handle r-alt in addition to existing l-alt support
Still not trying to make this DRY or anything; I just want to write it
out long-hand and see what patterns emerge.

Idea here is to track hardware modifier state (what keys have been
pressed on physical keyboard) and virtual modifier state (what keys
we've sent downstream to the virtual keyboard that X and apps are
actually seeing). This is a small step toward being able to handle more
complex combinations of modifiers. Once I get a feel for the pattern
I'll to make this all a little more elegant and concise.