feat(interception): add meta_press() and meta_release() functions
Not used yet, but probably will shortly. Note that the terminology here
is confusing.

Kernel events use "meta" for the "super" key. That's the sense in which
"meta" is used here.

But xkeysnail uses "M" and "Alt" interchangeably; the "m" there should
not be confused with "meta" in the kernel sense:

https://github.com/mooz/xkeysnail