feat(interception): make mapping definitions more expressive
Up until now we have effectively been embodying our state and the rules
for transitioning between states in imperative code, but the plan was
always to make this more declarative in nature, once I'd seen the basic
idea work.

So now we define our desired mappings in the form of C structs. I am not
going to write or import a YAML or JSON parser because the structs are
just fine in terms of readability.

Given those structs, I rewrote the code to process them somewhat
dynamically, and although I have only briefly tested it, it does seem to
provide the correct behavior. And as you can see, it also handles some
of the more "complicated" cases that we weren't addressing in the old
implementation (eg. Alt-[ and Alt-] to Ctrl-Shift-Tab and Ctrl-Tab
respectively). Other than that, I think I've brought everything across.

Test plan: Just in case I botch things, I test for 60 seconds at a time,
with an autorevert after that interval; eg:

    systemctl disable udevmon # ie. allow last-resort recovering by powercycling
    make
    sudo cmake --install . --prefix /usr
    sudo systemctl start udevmon; sleep 3; asdf; sleep 60; sudo systemctl stop udevmon
    xev -event keyboard