feat(interception): map Ctrl-I to F6
Ideally this would be a kitty-only thing (and in fact, I wonder if I can
bind it there). Note the limitation anyway that this only applies to the
physical control key (that we see at the libevdev level) and not the
physical caps lock key (that only looks like control to layers higher up
in the stack).