refactor(dotfiles): don't auto-assign obs to workspace 5
My terminal is usually on workspace 1, but obs can only meaningfully
stream windows that are on the workspace where it starts. So you
actually want to open it on workspace 1, set up the window capture, and
then move obs to another workspace.

You can also just share your entire screen, but there are obvious
privacy concerns with that, so caution is warranted.