fix(dotfiles): use /bin/zsh in shebang lines
Because on macOS I have:

    /bin/zsh - provided by Apple
    /usr/local/bin/zsh - via Homebrew

And on Arch:

    /usr/bin/zsh - provided by OS
    /bin/zsh - via symlink (because /bin is a link to /usr/bin)

So, `/bin/zsh` is going to work on both. Even though this means using an
older version on macOS (Apple is providing 5.3; vs 5.8 via
Homebrew/Arch) this should be fine because I am not using any fancy new
features in these scripts.