chore(dotfiles): update default branch name to "main"
Note, this won't actually do anything on macOS (at least, not on the
version I am currently using, which is Mojave), because it requires Git
v2.28 or later, and Apple ships v2.20.1. But, it will work on Linux, and
eventually work on macOS when Apple catches up.

    https://github.blog/2020-07-27-highlights-from-git-2-28/#introducing-init-defaultbranch

This isn't a change I was in a huge hurry to make anyway, because I
didn't want to break existing tools, links, and references. But it's
been over 6 months since people started talking about this, and there
has been time now for tool-makers to start making adjustments. Because
"main" is the GitHub default, we can expect to see "main" eventually
predominate over "master", at which point, _delaying_ the change further
is going to be the thing at risk of breaking things. As it is, it is a
fairly minimal change, affecting only newly created repositories, so I
don't think this is actually going to break anything at all.