feat(dotfiles): add irssi config
I'm using this to have Twitch chat visible in streams.

The `default.theme` is totally vanilla with only one change. I changed:

    msgnick = "%K<%n$0$1-%K>%n %|";

to:

    msgnick = "%K<%n$0$1-%K>%n ";

to avoid the excessive indenting. I have irssi off in a very narrow side
panel, so this indenting was crazy:

    < greghurrell> this is me
                   saying something
                   and getting
                   wrapped too much

This is better:

    < greghurrell> this is me
       saying something and getting
       wrapped too much

The two-space indent is achieved with:

    /set indent 2
    /save

which is reflected in the `config` file. Note: that file is encrypted
because it contains an OAuth token.

Indenting tip via:

    https://askubuntu.com/a/269441

Docs are scant, but I found this describing what `%|` does:

    https://github.com/irssi/irssi/blob/master/docs/formats.txt

In short, it "Marks the indentation position".