chore(dotfiles): make i3 look more minimalistic
After using the gaps settings for a while now, I feel pretty comfortable
concluding that I hate the way it looks.

So, hide the ugly bar (it will only show when Alt is pressed), and kill
all the gaps. Add in a 1 pixel border when there are multiple windows on
the screen to stop them from crashing into each other.