chore(vim): turn on g:completion_auto_change_source
Not sure whether this is the best idea or whether I should be adding a
plugin like:

    https://github.com/steelsojka/completion-buffers

Basically, I want it to complete words in buffers automatically (without
having to press C-n). So, giving this a try... it basically means that
if there are no matches from LSP or snippets, it will fall back and show
buffer options. I have a feeling that I'll end up going with the plugin
because this isn't a fuzzy search, but I still want to give this a try
before committing to yet another dependency.