refactor(vim): switch to nvim-compe
So, some perf problems with completion-nvim were pointed out to me here:

    https://twitter.com/ahmedelgabri/status/1397571165447671813

but here's the kicker:

    https://github.com/nvim-lua/completion-nvim/issues/357

It turns out that the "semi-official" status of completion-nvim by
virtue of being under the "nvim-lua" org is under question... it might
not be there for much longer.

nvim-compe seems a mostly drop-in replacement. The only thing that I can
obviously see that doesn't work out of the box is signature help:

    https://github.com/hrsh7th/nvim-compe/issues/191