chore: add host-specific Zsh config
In doing this, I noticed an error in my old alias: it needed to be
quoted. And adding the quotes uncovered a bug: when you do `gets` in
Ruby it is going to read from stdin by default only in the absence
of command line arguments; now that we are passing arguments, the
`gets` doesn't work any more. This is fixed by adding an explicit
`STDIN`.