fix: make Homebrew sort of work on M1 Mac
As noted here:

    https://github.com/wincent/wincent/issues/102

I can't really run the Homebrew installation script in an
automated fashion any more. First of all it fails to run
`sudo`. If you prime it by running `sudo` first, it will
run, but it will install into `/usr/local/` even though
`/opt/homebrew/` is the right location on M1 Macs.

So, I removed the install with the uninstall script:

    https://raw.githubusercontent.com/Homebrew/install/master/uninstall.sh

then I re-ran the install manually, and it went into
`/opt/homebrew/` as desired.

I need to figure out how to make this play nicely on both
old and new Macs in a robust/tidy way, but for now this
commit should be unobstrusive enough.

In a follow-up commit I will be commenting out some items
from the Brewfile that don't install properly on ARM.