fix(fonts): make sure fonts actually get copied
So "aspects/fonts/files/source-code-pro/" is actually a symlink to
somewhere under "vendor/", which means that our `isDirectory() ||
isFile()` check was skipping right over it and not actually copying the
files.

The fix, then, is to look at symlinks that we find during traversal and
`stat` the object they point to to see whether _that_ is a directory or
file, and then proceed.