chore(defaults): remove problematic defaults
On new Big Sur install, I got these errors:

    Could not write domain com.apple.universalaccess; exiting

For these items.

Now, the first one, "reduceTransparency", looks horrible on Big Sur, so
let's dump it.

Second one, "closeViewSmoothImages", seems to be the default anyway, so
let's kill.

Finally, the third, I can't even find this in the UI any more
because they have renamed it. The closest thing I could find was
"closeViewTrackpadGestureZoomEnabled" but it doesn't seem that useful,
so I'm just going to kill this.