fix(dotfiles): avoid ordering issues setting up $PATH
In new terminal windows `vim` wasn't getting aliased to `nvim`, but it
was in tmux sessions and nested shells.

This is because the alias is gated with an `if command -v nvim` check,
and we had a subtle `$PATH` ordering issue.

On my Intel machine, `/usr/local/bin` is in the `$PATH` and the
executable is found. On Apple Silicon, `nvim` is in `/opt/homebrew/bin`
and that doesn't get set up until too late.

So, I looked into this further and noticed some unfortunate circular
dependencies:

- `path` depended on `exports`, and vice versa
- `aliases` depended on `path`

The fix is to move `path` to the top, and pull the things in `exports`
that it depended on into the parent file. It's all very ugly, but
nothing else seems better (ie. we could just collapse everything into
one file and intermingle as though we were writing C without forward
declarations...)